+++ Source/PsapSimulator/Settings/AppSettings.cs
@@ -8,6 +8,7 @@ using SipLib.Logging;
 using System.Text.Json.Serialization;
 using SipRecClient;
 using SipLib.TestCalls;
+using System.IO;
 
 /// <summary>
 /// All settings for the application