+++ Source/PsapSimulator/Settings/TransferSettings.cs
@@ -4,6 +4,7 @@
 
 using Ng911Lib.Utilities;
 using SipLib.Logging;
+using System.IO;
 
 namespace PsapSimulator.Settings;