+++ Source/PsapSimulator/WindowsAudio/WindowsAudioUtils.cs
@@ -4,6 +4,7 @@
 
 using NAudio.Wave;
 using SipLib.Media;
+using System.IO;
 
 namespace WindowsWaveAudio;