+++ VersionHistory.md
@@ -1,5 +1,16 @@
 # Version History
 
+## v1.1.0 - TBD
+| Issue No. | Change Type | Description |
+|--------|--------|-------|
+| NA     | Change   | Changed the DisabledCamera.jpg file in the Recordings directory to a head and shoulders image. |
+| NA     | Addition | Added the Add Media button to the Current Call dialog. This button allows the user to add a new media type to the call. |
+| NA     | Change   | Added an a=record:on attribute to each media description if SIPREC is enabled. |
+| NA     | Fix      | Was polling configured and enabled NG9-1-1 Event Loggers even though NG9-1-1 Event Logging was disabled. |
+| NA     | Fix      | The number of enabled NG9-1-1 Event Loggers was not properly displayed in the Interfaces tab of the Settings form. |
+| NA     | Fix      | A SIPValidationException occurred in the NG9-1-1 Event Logger Settings form if the Logger URI setting was not a valid URI. |
+
+
 ## v1.0.0 - 25 Sep 2025
 | Issue No. | Change Type | Description |
 |--------|--------|-------|